backend/requirements.txt
google-auth-oauthlib
google-api-python-client
reportlab
python-docx
flask
